backend/requirements.txt
flask
python-binance
pandas
ta